Harden hosted BACnet services and concurrency.
Fixes a bug causing the server to crash due to malformed whois packets.
Correct hosted object semantics, and support bounded concurrent requests over one UDP socket.
Consolidate command priority handling, BACnet error representation, and response framing while preserving bacnet-stack interoperability.

+++ src/client/mod.rs
@@ -1099,8 +1099,8 @@ impl BacnetClient {
                 error_code,
                 ..
             } if invoke_id == expected_invoke_id => Err(ClientError::PropertyError {
-                class: error_class as u32,
-                code: error_code as u32,
+                class: error_class,
+                code: error_code,
             }),
             Apdu::Reject {
                 invoke_id,
